+++ packages/api/src/enums/webSocketEvent.enum.ts
@@ -1,5 +1,6 @@
 export enum WebSocketEvent {
   CHECK_DEVICE_CONNECTION = 'checkDeviceConnection',
   TOGGLE_GATE = 'toggleGate',
+  DEVICE_TURNED_ON = 'deviceTurnedOn',
   SET_NGROK_DATA = 'setNgrokData',
 }