+++ packages/api/src/modules/app.module.ts
@@ -8,6 +8,7 @@ import { ConfigModule } from './config/config.module';
 import { ConfigurationModule } from './configuration/configuration.module';
 import { DatabaseModule } from './database/database.module';
 import { ExternalIntegrationsModule } from './external-integrations/external-integrations.module';
+import { HistoryModule } from './history/history.module';
 import { InvitationsModule } from './invitations/invitations.module';
 import { MailerModule } from './mailer/mailer.module';
 import { PasswordResetModule } from './password-reset/password-reset.module';
@@ -43,6 +44,7 @@ import { WebsocketModule } from './websocket/websocket.module';
     ExternalIntegrationsModule,
     ConfigurationModule,
     CameraModule,
+    HistoryModule,
   ],
 })
 export class AppModule {}