+++ packages/api/src/modules/history/dto/create-history-device-event.dto.ts
@@ -1,0 +1,8 @@
+import { IsEnum } from 'class-validator';
+
+import { HistoryEvent } from '../../../enums/historyEvent.enum';
+
+export class CreateHistoryDeviceEventDto {
+  @IsEnum(HistoryEvent)
+  event: HistoryEvent.TurnedOn | HistoryEvent.TurnedOff;
+}