+++ packages/api/src/modules/repository/history.repository.ts
@@ -1,0 +1,4 @@
+import { HistoryEntity } from '../database/entities/history.entity';
+import { BaseRepository } from './base.repository';
+
+export class HistoryRepository extends BaseRepository(HistoryEntity) {}