+++ packages/api/src/modules/websocket/websocket.gateway.ts
@@ -2,6 +2,7 @@ import {
   CACHE_MANAGER,
   Inject,
   Injectable,
+  InternalServerErrorException,
   Logger,
   MethodNotAllowedException,
   ServiceUnavailableException,
@@ -17,10 +18,16 @@ import {
 import { Cache } from 'cache-manager';
 import { Server, Socket } from 'socket.io';
 
+import { HistoryEvent } from '../../enums/historyEvent.enum';
 import { WebSocketEvent } from '../../enums/webSocketEvent.enum';
 import { NgrokData } from '../camera/camera.controller';
+import { UserEntity } from '../database/entities/user.entity';
+import { HistoryService } from '../history/history.service';
 import { TicketService } from '../ticket/ticket.service';
+import { UsersService } from '../users/users.service';
 import { WebsocketConfigService } from './config/websocket-config.service';
+
+const HEROKU_RECONNECT_DELAY = 30;
 
 @Injectable()
 @WebSocketGateway({
@@ -33,6 +40,8 @@ export class Websocket implements OnGatewayInit, OnGatewayConnection, OnGatewayD
     @Inject(CACHE_MANAGER) private readonly cacheManager: Cache,
     private readonly ticketService: TicketService,
     private readonly websocketConfigService: WebsocketConfigService,
+    private readonly historyService: HistoryService,
+    private readonly usersService: UsersService,
   ) {}
 
   @WebSocketServer() server: Server;
@@ -70,12 +79,40 @@ export class Websocket implements OnGatewayInit, OnGatewayConnection, OnGatewayD
     this.logger.log(`Client connected: ${client.id} with userId: ${userId}`);
   }
 
+  async toggleGate(socketClient: undefined, user: UserEntity): Promise<void>;
+  async toggleGate(socketClient: Socket): Promise<void>;
   @SubscribeMessage(WebSocketEvent.TOGGLE_GATE)
-  toggleGate(): void {
+  async toggleGate(socketClient?: Socket, user?: UserEntity): Promise<void> {
+    if (socketClient && user) {
+      throw new MethodNotAllowedException();
+    }
+
+    if (socketClient) {
+      const userId = this.clients.get(socketClient.id);
+      if (!userId) {
+        throw new InternalServerErrorException('User not connected over websockets');
+      }
+
+      const connectedUser = await this.usersService.findOne(userId);
+      this.logger.log(`User toggle gate (id: ${connectedUser.id}, email: ${connectedUser.email})`);
+      await this.historyService.create({ event: HistoryEvent.Open, user: connectedUser });
+    }
+
+    if (user) {
+      this.logger.log(`User toggle gate (id: ${user.id}, email: ${user.email})`);
+      await this.historyService.create({ event: HistoryEvent.Open, user });
+    }
+
     if (!this.deviceClient) {
       throw new ServiceUnavailableException('Gate disconnected');
     }
     this.deviceClient.send(WebSocketEvent.TOGGLE_GATE);
+  }
+
+  @SubscribeMessage(WebSocketEvent.DEVICE_TURNED_ON)
+  private async deviceTurnedOn(): Promise<void> {
+    await this.historyService.create({ event: HistoryEvent.TurnedOn });
+    this.logger.log('Device turned on');
   }
 
   @SubscribeMessage(WebSocketEvent.SET_NGROK_DATA)
@@ -88,8 +125,8 @@ export class Websocket implements OnGatewayInit, OnGatewayConnection, OnGatewayD
   }
 
   @SubscribeMessage(WebSocketEvent.CHECK_DEVICE_CONNECTION)
-  checkDeviceConnection(client: Socket): void {
-    this.logger.log(`client:${this.clients.get(client.id) ?? ''}`);
+  checkDeviceConnection(): void {
+    this.logger.log('Check device connection');
     this.server.emit(WebSocketEvent.CHECK_DEVICE_CONNECTION, Boolean(this.deviceClient));
   }
 
@@ -98,6 +135,13 @@ export class Websocket implements OnGatewayInit, OnGatewayConnection, OnGatewayD
     if (this.deviceClient && client.id === this.deviceClient.id) {
       this.deviceClient = undefined;
       this.server.emit(WebSocketEvent.CHECK_DEVICE_CONNECTION, false);
+      // eslint-disable-next-line @typescript-eslint/no-misused-promises
+      setTimeout(async () => {
+        if (!this.deviceClient) {
+          this.logger.log('Device turned off');
+          await this.historyService.create({ event: HistoryEvent.TurnedOff });
+        }
+      }, HEROKU_RECONNECT_DELAY * 1000);
     } else {
       this.clients.delete(client.id);
     }