+++ packages/client/src/elements/buttons/index.ts
@@ -1,4 +1,5 @@
 export { default as BackButton } from './BackButton';
 export { default as Button } from './Button';
 export { default as CloseButton } from './CloseButton';
+export { default as DeleteHoverButton } from './DeleteHoverButton';
 export { default as IconButton } from './IconButton';