+++ packages/client/src/elements/fields/DateField/index.tsx
@@ -1,24 +1,36 @@
 import React from 'react';
-import { useTranslation } from 'react-i18next';
 
+import { useFormatDate } from '../../../hooks';
 import { BaseRecordField } from '../Fields.types';
 import { DateFieldProps } from './DateField.types';
 
-const DateField = <T extends BaseRecordField>({ source, record, showTime }: DateFieldProps<T>) => {
-  const { i18n } = useTranslation();
+const DateField = <T extends BaseRecordField>({
+  source,
+  record,
+  showTime,
+  style,
+}: DateFieldProps<T>) => {
+  const formatDate = useFormatDate();
 
   if (!record || !record[source]) {
-    return <p data-testid="dateField">-</p>;
+    return (
+      <p data-testid="dateField" style={style}>
+        -
+      </p>
+    );
   }
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-  const date = new Date(record[source].toString());
-  const formatter = new Intl.DateTimeFormat(i18n.language, {
+
+  const formatterOptions: Intl.DateTimeFormatOptions = {
     timeStyle: showTime ? 'medium' : undefined,
     dateStyle: 'short',
-  });
-  const formattedDate = formatter.format(date);
+  };
+  const dateLabel = formatDate(record[source], formatterOptions);
 
-  return <p data-testid="dateField">{formattedDate}</p>;
+  return (
+    <p data-testid="dateField" style={style}>
+      {dateLabel}
+    </p>
+  );
 };
 
 DateField.displayName = 'DateField';