+++ packages/client/src/elements/lists/HistoryCompactList/HistoryCompactList.constants.ts
@@ -1,0 +1,1 @@
+export const RECORD_ICON_CIRCLE_SIZE = '28px';