+++ packages/client/src/interfaces/utils.types.d.ts
@@ -1,0 +1,5 @@
+import { ITheme } from '../theme/Theme';
+
+type StyledHelperFunc<T> = T & {
+  theme: ITheme;
+};